% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{Geom prototype to draw the hurricane}
\format{An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomHurricane
}
\description{
Geom prototype to draw the hurricane
}
\keyword{datasets}
