% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{read_storm_data}
\alias{read_storm_data}
\title{Read storm observation from the `ebtrk_atlc_1988_2015.txt` file}
\usage{
read_storm_data(datadir = system.file("extdata", package =
  "CourseraRdataVisualizationFinal"))
}
\arguments{
\item{datadir}{Directory where the file is stored}
}
\value{
The tidy data.frame representing the storms

storm_id                date latitude longitude wind_speed  ne  nw  se
Katrina-2005 2005-08-29 12:00:00     29.5     -89.6         34 200 100 200
Katrina-2005 2005-08-29 12:00:00     29.5     -89.6         50 120  75 120
Katrina-2005 2005-08-29 12:00:00     29.5     -89.6         64  90  60  90
}
\description{
Read storm observation from the `ebtrk_atlc_1988_2015.txt` file
}
