% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{draw_katrina}
\alias{draw_katrina}
\title{Draw the Katrina hurricane}
\usage{
draw_katrina(datadir)
}
\arguments{
\item{datadir}{Directory where the file is stored}
}
\description{
Draw the Katrina hurricane
}
