% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{polygon_points}
\alias{polygon_points}
\title{Compute the points representing the storm polygon}
\usage{
polygon_points(x, y, ne, nw, se, sw, scale_radii = 1)
}
\arguments{
\item{x}{X coordinate of the center}

\item{y}{Y coordinate of the center}

\item{ne}{Radius of the north-east quarter}

\item{nw}{Radius of the north-west quarter}

\item{se}{Radius of the south-east quarter}

\item{sw}{Radius of the south-west quarter}

\item{scale_radii}{Scale of the polygon, default to 1}
}
\value{
The resulting coordinates (data.frame containing columns `x` and `y` for coordiantes)
}
\description{
Compute the points representing the storm polygon
}
